requirements/local.txt
# Local development dependencies go here
-r test.txt
Sphinx==1.2.2
ansible==1.6.6
django-extensions==1.3.8
pydot2==1.0.33
coverage==3.7.1
